feat: remove theme switch
It added much complexity for not much gain. It's best to rely on the
system-wide theme preference. This also allows me to purge ALL
JavaScript from the site, which is nice. Not that there was much JS to
begin with — the theme swithcer was purely HTML+CSS, and JS was only
used to cache the theme preference. Unnecessary complexity. Good
riddance. Tho, I will miss the cute little sun and moon icons...